completed landing page, added sort svg, fixed button highlight
